+++ changelog.md
@@ -4,6 +4,8 @@
 ### New Modules
 - Bounded VCO
 - DASE (Dynamic Amplifying Shaping Envelope)
+- Envelope Factory
+- Envelope Stage Expander
 - Linear Merge
 - Linear Merge Expander
 - Ratio